break apart runner
