+++ src/cfnlint/runner/deployment_files/__init__.py
@@ -1,0 +1,4 @@
+"""
+Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+SPDX-License-Identifier: MIT-0
+"""
